==> 0 - Lemmatization.ipynb <==
import pandas as pd
from preprocessing import preprocessing_df

### Preprocessing : Cleaning + Lemmatization

df = pd.read_json('data/mails_raw.json', encoding='utf-16')
df

df2 = preprocessing_df(df)
df2

df2 = df2.drop('index', axis = 1)

subset = list(df2.columns)
subset.remove('label')
subset.remove('from')
df3 = df2.drop_duplicates(subset = subset, ignore_index=True)
print(df2.shape, df3.shape)
df3.head()

df3.to_json('data/mails_preprocessing.json', force_ascii=False, orient = 'records')

sorted([word for sentence in df2.text_lem for word in sentence.split(' ')], key = lambda x:-len(x))
