==> 4.2 - LSTM + CNN (embeddings).ipynb <==
# Bigrams + Word embedding + RNN

import pandas as pd
import numpy as np
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.utils import compute_class_weight
from keras.models import Sequential
from keras.layers import Dense, LSTM, Dropout, LSTM, Conv1D, MaxPooling1D
from keras.callbacks import EarlyStopping
from sklearn.metrics import confusion_matrix, classification_report, f1_score
import keras.backend as K
np.random.seed(42)

data = pd.read_pickle('data/mails_embedded_doc2vec_bigrams.pkl')
data.head()

def custom_f1(y_true, y_pred):    
    def recall_m(y_true, y_pred):
        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        Positives = K.sum(K.round(K.clip(y_true, 0, 1)))
        
        recall = TP / (Positives+K.epsilon())    
        return recall 
    
    
    def precision_m(y_true, y_pred):
        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        Pred_Positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
    
        precision = TP / (Pred_Positives+K.epsilon())
        return precision 
    
    precision, recall = precision_m(y_true, y_pred), recall_m(y_true, y_pred)
    
    return 2*((precision*recall)/(precision+recall+K.epsilon()))

X_unlabeled = data[data._label_==2].drop('_label_', axis = 1)
X, y = data[data['_label_'] != 2].drop('_label_', axis = 1), data[data['_label_'] != 2]._label_
X, X_test, y, y_test = train_test_split(X.values, y.values, test_size=0.15, random_state = 42)
X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.15, random_state = 42)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_val = scaler.transform(X_val)
X_test = scaler.transform(X_val)
X_train = np.reshape(X_train, (X_train.shape[0], 1, X_train.shape[1]))
X_val = np.reshape(X_val, (X_val.shape[0], 1, X_val.shape[1]))
X_test = np.reshape(X_test, (X_test.shape[0], 1, X_test.shape[1]))

class_weight = compute_class_weight('balanced', classes = [0,1], y = y_train)
class_weight = {0:class_weight[0], 1:class_weight[1]}
class_weight

## LSTM

model1=Sequential()
model1.add(LSTM(128, input_shape = (1, 518), return_sequences=True))
model1.add(LSTM(64, return_sequences=True))
model1.add(Dense(256,activation='relu'))
model1.add(Dropout(0.5))
model1.add(Dense(8,activation='relu'))
model1.add(Dense(1,activation='sigmoid'))
model1.compile(loss='binary_crossentropy',optimizer='adam',metrics=[custom_f1])
print(model1.summary())

model1.fit(X_train,y_train,batch_size=16,epochs=10,
          validation_split=0.2,callbacks=[EarlyStopping(monitor='val_custom_f1',min_delta=0.000001)], class_weight=class_weight,
          workers = 6, use_multiprocessing = True)

y_prob = model1.predict(X_val)
y_pred = np.array([1*(pred>=0.5) for pred in y_prob])
y_pred = y_pred[:,0,0]

CM = confusion_matrix(y_val, y_pred)
print(CM)
print(classification_report(y_val, y_pred))

# CNN + LSTM

model3 = Sequential()
model3.add(Conv1D(filters=32, kernel_size=3, input_shape = (1, 518), padding='same', activation='relu'))
model3.add(MaxPooling1D(pool_size=2, padding = 'same'))
model3.add(LSTM(100))
model3.add(Dropout(0.5))
model3.add(Dense(8,activation='relu'))
model3.add(Dense(1, activation='sigmoid'))
model3.compile(loss='binary_crossentropy', optimizer='adam', metrics=[custom_f1])
print(model3.summary())

class_weight = compute_class_weight('balanced', classes = [0,1], y = y_train)
class_weight = {0:class_weight[0], 1:class_weight[1]}
class_weight

model3.fit(X_train,y_train, batch_size = 32,epochs=10,
          validation_split=0.2,callbacks=[EarlyStopping(monitor='val_loss',min_delta=0.0001)], class_weight=class_weight,
          workers = 6, use_multiprocessing = True)

y_pred = model3.predict(X_val)
y_pred = np.array([1*(pred>=0.5) for pred in y_pred])
y_pred = y_pred[:,0]

CM = confusion_matrix(y_val, y_pred)
print(CM)
print(classification_report(y_val, y_pred))

# Subsampling

x1 = data[data._label_==0].sample(590, random_state = 42)
y1 = x1._label_
x1 = x1.drop('_label_', axis = 1)
x2 = data[data._label_==1].sample(590, random_state = 42)
y2 = x2._label_
x2 = x2.drop('_label_', axis = 1)
Xt = np.concatenate([x1,x2])
yt = np.concatenate([y1, y2])

Xt_train, Xt_test, yt_train, yt_test = train_test_split(Xt, yt, test_size=0.15, random_state=42)
Xt_train, Xt_val,yt_train, yt_val = train_test_split(Xt_train, yt_train, test_size=0.15, random_state=42)

scaler = StandardScaler()
Xt_train = scaler.fit_transform(Xt_train)
Xt_val = scaler.transform(Xt_val)
Xt_test = scaler.transform(Xt_val)
Xt_train = np.reshape(Xt_train, (Xt_train.shape[0], 1, Xt_train.shape[1]))
Xt_val = np.reshape(Xt_val, (Xt_val.shape[0], 1, Xt_val.shape[1]))
Xt_test = np.reshape(Xt_test, (Xt_test.shape[0], 1, Xt_test.shape[1]))

model1=Sequential()
model1.add(LSTM(128, input_shape = (1, 518), return_sequences=True))
model1.add(LSTM(64, return_sequences=True))
model1.add(Dense(256,activation='relu'))
model1.add(Dropout(0.5))
model1.add(Dense(8,activation='relu'))
model1.add(Dense(1,activation='sigmoid'))
model1.compile(loss='binary_crossentropy',optimizer='adam',metrics=[custom_f1])
print(model1.summary())

model1.fit(Xt_train,yt_train,batch_size=16,epochs=100,
          validation_split=0.2, verbose = 0, workers = 6, use_multiprocessing = True)

y_prob = model1.predict(Xt_val)
y_pred = np.array([1*(pred>=0.5) for pred in y_prob])
y_pred = y_pred[:,0,0]

CM = confusion_matrix(yt_val, y_pred)
print(CM)
print(classification_report(yt_val, y_pred))

model3 = Sequential()
model3.add(Conv1D(filters=32, kernel_size=3, input_shape = (1, 518), padding='same', activation='relu'))
model3.add(MaxPooling1D(pool_size=2, padding = 'same'))
model3.add(LSTM(100))
model3.add(Dropout(0.5))
model3.add(Dense(8,activation='relu'))
model3.add(Dense(1, activation='sigmoid'))
model3.compile(loss='binary_crossentropy', optimizer='adam', metrics=[custom_f1])
print(model3.summary())

model3.fit(Xt_train,yt_train, batch_size = 16,epochs=100,
          validation_split=0.2,verbose = 0, workers = 6, use_multiprocessing = True)

y_pred = model3.predict(Xt_val)
y_pred = np.array([1*(pred>=0.5) for pred in y_pred])
y_pred = y_pred[:,0]

CM = confusion_matrix(yt_val, y_pred)
print(CM)
print(classification_report(yt_val, y_pred))

def create_model():
    model = Sequential()
    model.add(Conv1D(filters=32, kernel_size=3, input_shape = (1, 518), padding='same', activation='relu'))
    model.add(MaxPooling1D(pool_size=2, padding = 'same'))
    model.add(LSTM(100))
    model.add(Dropout(0.5))
    model.add(Dense(8,activation='relu'))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy', custom_f1])
    return model

from keras.wrappers.scikit_learn import KerasClassifier
from sklearn.model_selection import StratifiedKFold, cross_val_score
nn_model = KerasClassifier(build_fn=create_model, epochs = 15, batch_size = 32, verbose = 0)

kfold = StratifiedKFold(n_splits=10, shuffle=True)
results = cross_val_score(nn_model, Xt_train, yt_train, cv=kfold, scoring = 'f1', verbose = 0)

results.mean()
