==> 6 - CNN + Bi-LSTM + meta.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
from collections import Counter
from sklearn.model_selection import ShuffleSplit, train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.utils import compute_class_weight
from tensorflow import keras
from tensorflow.keras import layers
from sklearn.metrics import confusion_matrix, classification_report, f1_score
import keras.backend as K
import joblib
np.random.seed(42)

data = pd.read_pickle('data/mails_embedded_doc2vec_bigrams.pkl')
data.head()

# subsampling data
np.random.seed(42)
df1 = data[data._label_==0].sample(590)
df2 = data[data._label_==1].sample(590)
df_subsampled = pd.concat([df1,df2], ignore_index=True)

# no subsampling
df = data[data._label_!=2]

shuffler = ShuffleSplit(n_splits=1,test_size=0.3, random_state=42)
train_indexes, test_indexes = list(shuffler.split(df.index))[0]
y_train = df.iloc[train_indexes,:]._label_
df_train = df.iloc[train_indexes,:].drop('_label_', axis = 1)
y_test = df.iloc[test_indexes,:]._label_
df_test = df.iloc[test_indexes,:].drop('_label_', axis = 1)
print(Counter(y_train))

df_features_train = df_train.iloc[:,:18] #meta-data
df_features_test = df_test.iloc[:,:18]
df_embeddings_train = df_train.iloc[:,18:] #word embeddings
df_embeddings_test = df_test.iloc[:,18:]

class_weight = compute_class_weight('balanced', classes = [0,1], y = y_train)
class_weight = {0:class_weight[0], 1:class_weight[1]}
class_weight

features_scaler = StandardScaler()
X_features_train = features_scaler.fit_transform(df_features_train)
X_features_test = features_scaler.transform(df_features_test)
embedding_scaler = StandardScaler()
X_embeddings_train = embedding_scaler.fit_transform(df_embeddings_train)
X_embeddings_test = embedding_scaler.transform(df_embeddings_test)

X_features_train = np.reshape(X_features_train, (X_features_train.shape[0], 1, X_features_train.shape[1]))
X_features_test = np.reshape(X_features_test, (X_features_test.shape[0], 1, X_features_test.shape[1]))
X_embeddings_train = np.reshape(X_embeddings_train, (X_embeddings_train.shape[0], 1, X_embeddings_train.shape[1]))
X_embeddings_test = np.reshape(X_embeddings_test, (X_embeddings_test.shape[0], 1, X_embeddings_test.shape[1]))

def custom_f1(y_true, y_pred):    
    def recall_m(y_true, y_pred):
        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        Positives = K.sum(K.round(K.clip(y_true, 0, 1)))
        
        recall = TP / (Positives+K.epsilon())    
        return recall 
    
    
    def precision_m(y_true, y_pred):
        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
        Pred_Positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
    
        precision = TP / (Pred_Positives+K.epsilon())
        return precision 
    
    precision, recall = precision_m(y_true, y_pred), recall_m(y_true, y_pred)
    
    return 2*((precision*recall)/(precision+recall+K.epsilon()))

# Bi-LSTM

text_in = keras.Input(shape=(1,500,))
meta_in = keras.Input(shape=(1,18,))
lstm1 = layers.LSTM(100, return_sequences=True)(text_in)
lstm2 = layers.LSTM(100, go_backwards=True, return_sequences=True)(lstm1)
# lstm1 = layers.GRU(100, return_sequences = True)(text_in)
# lstm2 = layers.GRU(100, go_backwards=True, return_sequences = True)(lstm1)
dense1 = layers.Dense(256,activation='relu')(lstm2)
drop = layers.Dropout(0.5)(dense1)
merged = layers.concatenate([drop, meta_in])
dense1 = layers.Dense(64, activation='relu')(merged)
dense2 = layers.Dense(8, activation='relu')(dense1)
text_class = layers.Dense(2, activation='softmax')(dense2)
model1 = keras.Model([text_in, meta_in], text_class)

model1.compile(loss = 'categorical_crossentropy', optimizer = 'adam', metrics=[custom_f1])

from tensorflow.keras.utils import to_categorical
y_train_probas = to_categorical(y_train)
y_train_probas = np.reshape(y_train_probas, (y_train_probas.shape[0], 1, y_train_probas.shape[1]))

model1.fit([X_embeddings_train, X_features_train], y_train_probas, batch_size=32,
    epochs=200, validation_split=0.2, verbose = 0, workers = 6, use_multiprocessing = True)

y_prob = model1.predict([X_embeddings_test, X_features_test])
y_pred = y_prob.argmax(axis = -1)
y_pred[:,0]

CM = confusion_matrix(y_test, y_pred)
print(CM)
print(classification_report(y_test, y_pred))

# CNN + Bi-LSTM

text_in = keras.Input(shape=(1,500,))
meta_in = keras.Input(shape=(1,18,))
conv = layers.Conv1D(filters=32, kernel_size=3, input_shape = (1, 500), padding='same', activation='relu')(text_in)
max_pool = layers.MaxPooling1D(pool_size=2, padding = 'same')(conv)
lstm1 = layers.LSTM(100, return_sequences=True)(max_pool)
lstm2 = layers.LSTM(100, go_backwards=True, return_sequences=True)(lstm1)
dense1 = layers.Dense(256,activation='relu')(lstm2)
drop = layers.Dropout(0.5)(dense1)
merged = layers.concatenate([drop, meta_in])
dense2 = layers.Dense(64, activation='relu')(merged)
dense3 = layers.Dense(8, activation='relu')(dense2)
text_class = layers.Dense(1, activation='sigmoid')(dense3)
model2 = keras.Model([text_in, meta_in], text_class)

model2.compile(loss = 'binary_crossentropy', optimizer = 'adam', metrics=[custom_f1])

model2.fit([X_embeddings_train, X_features_train], y_train, batch_size=32,
    epochs=200, validation_split=0.2, verbose = 0, workers = 6, use_multiprocessing = True)

y_prob = model2.predict([X_embeddings_test, X_features_test])
y_pred = np.array([1*(pred>=0.5) for pred in y_prob])
y_pred = y_pred[:,0,0]

CM = confusion_matrix(y_test, y_pred)
print(CM)
print(classification_report(y_test, y_pred))

import joblib
# joblib.dump(features_scaler, 'models/features_scaler')
# joblib.dump(embedding_scaler, 'models/embedding_scaler')
model1.save('models/BILSTM_518features_input_emb_meta')
# model2.save('models/CNN_BILSTM_518features_input_emb_meta')
