==> 7 - Melusine.ipynb <==
from melusine.nlp_tools.embedding import Embedding
import pandas as pd
from melusine.models.neural_architectures import cnn_model, rnn_model, transformers_model, bert_model
from melusine.models.train import NeuralModel
from sklearn.metrics import classification_report
from sklearn.model_selection import ShuffleSplit, cross_validate
import numpy as np
from collections import Counter
from gensim.models.keyedvectors import KeyedVectors
import warnings
import joblib
np.random.seed(42)
warnings.simplefilter('ignore')

data = pd.read_json('data/mails_preprocessing.json').drop('from', axis = 1)
data.head()

embedding = Embedding(input_column='text_lem', min_count=2, method = 'word2vec_sg', size = 500)
embedding.train(data.astype({'text_lem':str}))
# embedding = pd.read_pickle('data/mails_embedded_doc2vec_bigrams.pkl')

embedding.save('models/MELUSINE_EMBEDDING')

features = [data.columns[3]] + list(data.columns[5:])
Counter(data.label)

# subsampling data
np.random.seed(42)
df1 = data[data.label==0].sample(590)
df2 = data[data.label==1].sample(590)
df_subsampled = pd.concat([df1,df2], ignore_index=True)

# no subsampling
df = data[data.label!=2]

shuffler = ShuffleSplit(n_splits=1,test_size=0.2, random_state=42)
train_indexes, test_indexes = list(shuffler.split(df.index))[0]
df_train = df.iloc[train_indexes,:]
df_test = df.iloc[test_indexes,:]
Counter(df_train.label)

# CNN

np.random.seed(42)
cnn_nn = NeuralModel(
    architecture_function = cnn_model,
    pretrained_embedding = embedding,
    text_input_column = 'text_lem',
    meta_input_list = features,
    batch_size = 32,
    embedding_dim = 500,
    seq_size=518,
    n_epochs=200
)

cnn_nn.fit(df_train, df_train.label, verbose = 0, workers = 6, use_multiprocessing = True)

y_pred = cnn_nn.predict(df_test)
print(classification_report(df_test.label, y_pred))

cnn_nn.save_nn_model('models/CNN_518features_input_text_lem')
joblib.dump(cnn_nn, 'models/CNN_518features_input_text_lem.pkl', compress=True)

# RNN

np.random.seed(42)
rnn_nn = NeuralModel(architecture_function=rnn_model,
    pretrained_embedding=embedding,
    text_input_column='text_lem',
    meta_input_list = features,
    batch_size = 32,
    embedding_dim = 500,
    seq_size=518,
    n_epochs=200
)

rnn_nn.fit(df_train, df_train.label, verbose = 0, workers = 6, use_multiprocessing = True)

y_pred = rnn_nn.predict(df_test)
print(classification_report(df_test.label, y_pred))

rnn_nn.save_nn_model('models/RNN_518features_input_text_lem')
joblib.dump(rnn_nn, 'models/RNN_518features_input_text_lem.pkl', compress=True)

# Transformer

np.random.seed(42)
transf_nn = NeuralModel(architecture_function=transformers_model,
    pretrained_embedding=embedding,
    text_input_column='text_lem',
    meta_input_list = features,
    batch_size = 32,
    embedding_dim = 500,
    seq_size=518,
    n_epochs=200
)

transf_nn.fit(df_train, df_train.label, verbose = 0, workers = 6, use_multiprocessing = True)

y_pred = transf_nn.predict(df_test)
print(classification_report(df_test.label, y_pred))

transf_nn.save_nn_model('models/TRANSF_518features_input_text_lem')
joblib.dump(transf_nn, 'models/TRANSF_518features_input_text_lem.pkl', compress=True)
